[sandbox] Use 32-bit offset accumulation in VisitFlat and IsEqualTo
This resolves a similar bugs as crrev.com/c/7926028, but located within
C++ flat string traversal (`String::VisitFlat`) and string comparison
(`String::IsEqualToImpl`). Under memory corruption, chaining sliced
strings previously allowed 32-bit signed slice offset accumulation to
overflow or underflow, enabling out-of-sandbox string reads.

This CL applies the same 32-bit unsigned wrapping arithmetic fix used in
Torque and Turboshaft (crrev.com/c/7926028) to both C++ locations,
ensuring that corrupted accumulated offsets wrap around safely within
4 GB and land inside the ExternalStringsCage or its guard regions.

TAG=agy
CONV=7b491e08-7ad2-40b6-804d-203ae21bd3f1

Bug: 524832525, 523940239
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7958979
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#108136}

+++ src/objects/string-inl.h
@@ -740,7 +740,7 @@ bool String::IsEqualToImpl(
 
   DisallowGarbageCollection no_gc;
 
-  int slice_offset = 0;
+  uint32_t slice_offset = 0;
   Tagged<String> string = this;
   const Char* data = str.data();
   while (true) {
@@ -760,7 +760,7 @@ bool String::IsEqualToImpl(
           return CompareCharsEqual(s->GetChars() + slice_offset, data, len);
         },
         [&](Tagged<SlicedString> s) {
-          slice_offset += s->offset();
+          slice_offset += static_cast<uint32_t>(s->offset());
           string = s->parent();
           return std::nullopt;
         },
@@ -1187,7 +1187,8 @@ Tagged<ConsString> String::VisitFlat(
     Visitor* visitor, Tagged<String> string, const int offset,
     const SharedStringAccessGuardIfNeeded& access_guard) {
   DisallowGarbageCollection no_gc;
-  int slice_offset = offset;
+  // TODO(524832525): Consider changing the offset parameter type to uint32_t.
+  uint32_t slice_offset = static_cast<uint32_t>(offset);
   const uint32_t length = string->length();
   DCHECK_LE(offset, length);
   while (true) {
@@ -1216,7 +1217,7 @@ Tagged<ConsString> String::VisitFlat(
               return Tagged<ConsString>();
             },
             [&](Tagged<SlicedString> s) {
-              slice_offset += s->offset();
+              slice_offset += static_cast<uint32_t>(s->offset());
               string = s->parent();
               return std::nullopt;
             },
